Consolidate copies of common_definition_helper
