+++ reaper-adm-extension/src/reaper_adm/CMakeLists.txt
@@ -6,6 +6,8 @@ add_library(Reaper_adm::reaper_adm ALIAS reaper_adm)
 find_package(Boost REQUIRED QUIET COMPONENTS filesystem)
 
 set(EXTENSION_SOURCES
+	${EPS_SHARED_DIR}/helper/common_definition_helper.cpp
+	
 	actionmanager.cpp
 	admchannel.cpp
 	admextraction.cpp
@@ -21,7 +23,6 @@ set(EXTENSION_SOURCES
 	channelrouter.cpp
 	color.cpp
 	commontrackpool.cpp
-	common_definition_helper.cpp
 	communicators.cpp
 	directspeakerautomationelement.cpp
 	elementcomparator.cpp
@@ -73,6 +74,7 @@ set(EXTENSION_SOURCES
 	)
 	
 set(EXTENSION_HEADERS
+	${EPS_SHARED_DIR}/helper/common_definition_helper.h
     ${EPS_SHARED_DIR}/helper/nng_wrappers.h
 	
 	actionmanager.h
@@ -91,7 +93,6 @@ set(EXTENSION_HEADERS
 	channelrouter.h
 	color.h
 	commontrackpool.h
-	common_definition_helper.h
 	communicators.h
 	directspeakerautomationelement.h
 	elementcomparator.h
